:art: :arrow_up: Refactor structure and upgrade to v0.24.13
- Renamed 'ca-gip' folder to 'cagip'
- Updated GroupName by removing '-' to avoid code-gen issues
- Generated new code using Kubernetes codegen 0.24.13

BREAKING CHANGE:
Imports can now be done via 'github.com/ca-gip/kubi/pkg/apis/cagip/v1'
