+++ deployments/kube-crds.yml
@@ -1,148 +1,141 @@
 ---
-apiVersion: apiextensions.k8s.io/v1beta1
+apiVersion: apiextensions.k8s.io/v1
 kind: CustomResourceDefinition
 metadata:
-  name: networkpolicyconfigs.ca-gip.github.com
+  name: networkpolicyconfigs.cagip.github.com
 spec:
-  group: ca-gip.github.com
+  group: cagip.github.com
   versions:
     - name: v1
       served: true
       storage: true
+      schema:
+        openAPIV3Schema:
+          type: object
+          properties:
+            spec:
+              type: object
+              properties:
+                egress:
+                  type: object
+                  properties:
+                    ports:
+                      type: array
+                      items:
+                        type: integer
+                        minimum: 1
+                        maximum: 65535
+                    cidrs:
+                      type: array
+                      items:
+                        type: string
+                        pattern: '^(\d{1,3}\.){3}\d{1,3}\/\d{1,2}$'
+                    namespaces:
+                      type: array
+                      items:
+                        type: string
+                ingress:
+                  type: object
+                  properties:
+                    namespaces:
+                      type: array
+                      items:
+                        type: string
+      additionalPrinterColumns:
+        - name: Egress-Ports
+          type: string
+          description: List of authorized port for egress traffic
+          jsonPath: .spec.egress.ports
+        - name: Egress-Cidr
+          type: string
+          description: List of authorized cidr for egress traffic
+          jsonPath: .spec.egress.cidrs
+        - name: Egress-Namespaces
+          type: string
+          description: List of egress namespaces
+          jsonPath: .spec.egress.namespaces
+        - name: Ingress-Namespaces
+          type: string
+          description: List of ingress namespaces
+          jsonPath: .spec.ingress.namespaces
   scope: Cluster
   names:
     plural: networkpolicyconfigs
     singular: networkpolicyconfig
     kind: NetworkPolicyConfig
     shortNames:
-      - netpolconf
-  additionalPrinterColumns:
-    - name: Egress-Ports
-      type: string
-      description: List of authorized port for egress traffic
-      JSONPath: .spec.egress.ports
-    - name: Egress-Cidr
-      type: string
-      description: List of authorized cidr for egress traffic
-      JSONPath: .spec.egress.cidrs
-    - name: Egress-Namespaces
-      type: string
-      description: List of egress namespaces
-      JSONPath: .spec.egress.namespaces
-    - name: Ingress-Namespaces
-      type: string
-      description: List of ingress namespaces
-      JSONPath: .spec.ingress.namespaces
-  validation:
-    # openAPIV3Schema is the schema for validating custom objects.
-    openAPIV3Schema:
-      properties:
-        spec:
-          type: object
-          properties:
-            egress:
-              type: object
-              properties:
-                ports:
-                  type: array
-                  items:
-                    type: integer
-                    minimum: 1
-                    maximum: 65535
-                cidrs:
-                  type: array
-                  items:
-                    type: string
-                    pattern: '^(\d{1,3}\.){3}\d{1,3}\/\d{1,2}$'
-                namespaces:
-                  type: array
-                  items:
-                    type: string
-            ingress:
-              type: object
-              properties:
-                namespaces:
-                  type: array
-                  items:
-                    type: string
-
+    - netpolconf
 ---
-apiVersion: apiextensions.k8s.io/v1beta1
+apiVersion: apiextensions.k8s.io/v1
 kind: CustomResourceDefinition
 metadata:
-  name: projects.ca-gip.github.com
+  name: projects.cagip.github.com
 spec:
-  group: ca-gip.github.com
+  group: cagip.github.com
   versions:
     - name: v1
       served: true
       storage: true
+      schema:
+        openAPIV3Schema:
+          type: object
+          properties:
+            spec:
+              type: object
+              required:
+              - project
+              properties:
+                whitelistAddresses:
+                  type: array
+                  items:
+                    type: string
+                tenant:
+                  type: string
+                environment:
+                  type: string
+                  enum:
+                  - development
+                  - integration
+                  - production
+                  - preproduction
+                  - uat
+                project:
+                  type: string
+                stages:
+                  type: array
+                  items:
+                    type: string
+                    enum:
+                    - scratch
+                    - staging
+                    - stable
+                sourceEntity:
+                  type: string
+      additionalPrinterColumns:
+        - name: Tenant
+          type: string
+          description: The customer aka company owning the cluster
+          jsonPath: .spec.tenant
+        - name: Environment
+          type: string
+          description: The environment for this namespace
+          jsonPath: .spec.environment
+        - name: Project
+          type: string
+          description: The project name ( without environment )
+          jsonPath: .spec.project
+        - name: Stages
+          type: string
+          description: Stages allowed for this project, can be mapped to registry or CI/CD
+          jsonPath: .spec.stages
+        - name: Source Entity
+          type: string
+          description: The refering entity ( LDAP group for the moment )
+          jsonPath: .spec.sourceEntity
   scope: Cluster
   names:
     plural: projects
     singular: project
     kind: Project
     shortNames:
-      - project
-  additionalPrinterColumns:
-    - name: Tenant
-      type: string
-      description: The customer aka company owning the cluster
-      JSONPath: .spec.tenant
-    - name: Environment
-      type: string
-      description: The environment for this namespace
-      JSONPath: .spec.environment
-    - name: Project
-      type: string
-      description: The project name ( without environment )
-      JSONPath: .spec.project
-    - name: Stages
-      type: string
-      description: Stages allowed for this project, can be mapped to registry or CI/CD
-      JSONPath: .spec.stages
-    - name: Source Entity
-      type: string
-      description: The refering entity ( LDAP group for the moment )
-      JSONPath: .spec.sourceEntity
-    - name: Contact
-      type: string
-      description: The refering contact for operation
-      JSONPath: .spec.contact
-  validation:
-    # openAPIV3Schema is the schema for validating custom objects.
-    openAPIV3Schema:
-      properties:
-        spec:
-          type: object
-          required:
-            - project
-          properties:
-            whitelistAddresses:
-              type: array
-              items:
-                type: string
-            tenant:
-              type: string
-            environment:
-              type: string
-              enum:
-                - development
-                - integration
-                - uat
-                - preproduction
-                - production
-            project:
-              type: string
-            stages:
-              type: array
-              items:
-                type: string
-                enum:
-                  - scratch
-                  - staging
-                  - stable
-            sourceEntity:
-              type: string
-            contact:
-              type: string
+    - project
