+++ hack/update-codegen.sh
@@ -5,7 +5,10 @@ set -o nounset
 set -o pipefail
 
 SCRIPT_ROOT=$(dirname "${BASH_SOURCE[0]}")/..
-CODEGEN_PKG=${CODEGEN_PKG:-$(cd "${SCRIPT_ROOT}"; ls -d -1 ./vendor/k8s.io/code-generator 2>/dev/null || echo ../code-generator)}
+CODEGEN_PKG=${CODEGEN_PKG:-$(
+  cd "${SCRIPT_ROOT}"
+  ls -d -1 ./vendor/k8s.io/code-generator 2>/dev/null || echo ../code-generator
+)}
 
 # generate the code with:
 # --output-base    because this script should also be able to run inside the vendor dir of
@@ -13,4 +16,4 @@ CODEGEN_PKG=${CODEGEN_PKG:-$(cd "${SCRIPT_ROOT}"; ls -d -1 ./vendor/k8s.io/code-
 #                  instead of the $GOPATH directly. For normal projects this can be dropped.
 bash "${CODEGEN_PKG}"/generate-groups.sh "informer,deepcopy,client,lister" \
   github.com/ca-gip/kubi/pkg/generated github.com/ca-gip/kubi/pkg/apis \
-  ca-gip:v1  --go-header-file "${SCRIPT_ROOT}"/hack/boilerplate.go.txt
+  cagip:v1 --go-header-file "${SCRIPT_ROOT}"/hack/boilerplate.go.txt
