--- pkg/apis/ca-gip/register.go
@@ -1,5 +1,0 @@
-package ca_gip
-
-const (
-	GroupName = "ca-gip.github.com"
-)
