+++ pkg/apis/cagip/register.go
@@ -1,0 +1,5 @@
+package cagip
+
+const (
+	GroupName = "cagip.github.com"
+)
